feat(spotify): add Spotify APIs
Add the Spotify router to the app.  The router has been updated as
previously users would be directed to the API and served with HTML to
confirm the connection is complete, this is now implemented as part of
the frontend which sends an API request to complete the connection.
This removes the very basic "Spotify connected successfully" page which
was used in the previous version.