feat(ci): update GitLab CI to build deployment zip
Add a build stage to GitLab CI for creating a deployment zip file. The
configuration now uses a Python 3.13 image, installs dependencies, and
packages code and dependencies into a zip artifact for deployment as a
Lambda function on AWS.  The CDK will download this file and perform
the actual deployment.